Rename config_ops -> dict_ops
