+++ manimlib/__init__.py
@@ -57,7 +57,7 @@ from manimlib.scene.scene import *
 
 from manimlib.utils.bezier import *
 from manimlib.utils.color import *
-from manimlib.utils.config_ops import *
+from manimlib.utils.dict_ops import *
 from manimlib.utils.customization import *
 from manimlib.utils.debug import *
 from manimlib.utils.directories import *
